1. Moved stop gdb server to separate function. 2. Added timeout after killing gdbserver. 3. Added test in integration-tests for stop gdb server.

+++ src/GDBTargetDebugSession.ts
@@ -234,8 +234,13 @@ export class GDBTargetDebugSession extends GDBDebugSession {
                 throw new Error('Missing stderr in spawned gdbserver');
             }
 
-            this.gdbserver.on('exit', (code) => {
-                const exitmsg = `${serverExe} has exited with code ${code}`;
+            this.gdbserver.on('exit', (code, signal) => {
+                let exitmsg: string;
+                if (code === null) {
+                    exitmsg = `${serverExe} is killed by signal ${signal}`;
+                } else {
+                    exitmsg = `${serverExe} has exited with code ${code}`;
+                }
                 this.sendEvent(new OutputEvent(exitmsg, 'server'));
                 if (!gdbserverStartupResolved) {
                     gdbserverStartupResolved = true;
@@ -342,13 +347,29 @@ export class GDBTargetDebugSession extends GDBDebugSession {
         }
     }
 
+    protected async stopGDBServer(
+    ): Promise<void> {
+        return new Promise((resolve, reject) => {
+            if (! this.gdbserver || this.gdbserver.exitCode !== null) {
+                resolve();
+            } else {
+                this.gdbserver.on('exit', () => {
+                    resolve();
+                });
+                this.gdbserver?.kill();
+            }
+            setTimeout(() => {reject()}, 1000);
+        });
+    }
+
     protected async disconnectRequest(
         response: DebugProtocol.DisconnectResponse,
         _args: DebugProtocol.DisconnectArguments
     ): Promise<void> {
         try {
             await this.gdb.sendGDBExit();
-            await this.gdbserver?.kill();
+            await this.stopGDBServer();
+            this.sendEvent(new OutputEvent('gdbserver stopped', 'server'));
             this.sendResponse(response);
         } catch (err) {
             this.sendErrorResponse(
